+++ src/KardFinancial/Commons/Types/OfferComponents.cs
@@ -1,9 +1,8 @@
 using global::System.Text.Json;
 using global::System.Text.Json.Serialization;
-using KardFinancial;
 using KardFinancial.Core;
 
-namespace KardFinancial.Users;
+namespace KardFinancial;
 
 /// <summary>
 /// UI component data for rendering offer details